[ICU-98] These are changes to support 64-bit platforms, such as Sun SPARC V9, which is the only one tested so far. X-Commit-URL: https://ssl.icu-project.org/trac/changeset/1022

+++ icu4c/source/common/ucnv2022.c
@@ -688,7 +688,7 @@ static const UConverterImpl _ISO2022Impl={
 };
 
 extern const UConverterSharedData _ISO2022Data={
-    sizeof(UConverterSharedData), ~0,
+    sizeof(UConverterSharedData), ~((uint32_t) 0),
     NULL, NULL, &_ISO2022Impl, "ISO_2022",
     2022, UCNV_IBM, UCNV_ISO_2022, 1, 4,
     { 0, 1, 0x1a, 0, 0, 0 }
